+++ LaserAPI/Models/Dto/Patterns/PointDto.cs
@@ -6,8 +6,8 @@ namespace LaserAPI.Models.Dto.Patterns
     {
         public Guid Uuid { get; set; }
         public Guid PatternUuid { get; set; }
-        public short X { get; set; }
-        public short Y { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
         public Guid ConnectedToUuid { get; set; }
     }
 }
